Letting send silent pushes
I'm not able to test GCM but APNS. So, I would like to allow sending silent pushes for APNS.  Because APNS (or iOS itself) sets default push sound if the `sound` property defined. Even you set an empty string or non-existing sound in the payload for the purpose of sending silent push.